+++ lib/glimmer/ui.rb
@@ -1,0 +1,5 @@
+module Glimmer
+  module UI
+  end
+  GUI = UI # alias
+end