+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Add proper indentation in code_text upon hitting ENTER
 - Reset Canvas Shape DSL alpha value to 255 when not explicitly set on a shape (Apply in Hello, Canvas! Sample)
 - Provide terse syntax for building canvas objects (autodetecting its width and height)
+- Provide terse syntax for building `:image_double_buffered` canvas objects (autodetecting its width and height):
 - Fix use of on_events in code_text widget with lines mode true
 
 ### 4.18.4.10